pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Template for backend API core project structure and objects"
authors = ["Oleksandr Besan <oleksandr.besan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
uvicorn = "^0.28.0"
loguru = "^0.7.2"
pydantic-settings = "^2.2.1"  
requests = "^2.31.0"  
asgi-correlation-id = "^4.3.1"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.9.0"
flake8-mypy = "^17.8.0"
flake8-type-checking = "^2.9.0"
types-requests = "^2.31.0.20240311"
pre-commit-hooks = "^4.5.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]